Update README.md
This project is a dynamic education platform offering courses, internships, and personalized profiles. It features responsive design, user authentication, video tutorials, and internship opportunities in IT fields like web development, data science, and more. With intuitive navigation and interactive features, it provides a seamless learning experience for students and professionals.